% Generated by build script
\name{cms_datasets}
\alias{cms_datasets}
\title{List all CMS provider data datasets}
\usage{
cms_datasets()
}
\arguments{
\item{max_results}{Max datasets to return (default: all)}
}
\value{tibble: identifier (dataset ID), title, description, modified}
\description{List all CMS provider data datasets}
