% Generated by build script
\name{cms_dialysis}
\alias{cms_dialysis}
\title{Dialysis facility information}
\usage{
cms_dialysis()
}
\arguments{
\item{state}{2-letter state code}
\item{max_results}{Max results (default 1000)}
}
\value{tibble with facility info and quality measures}
\description{Dialysis facility information}
