% Generated by build script
\name{cms_hospitals}
\alias{cms_hospitals}
\title{Hospital general information}
\usage{
cms_hospitals()
}
\arguments{
\item{state}{2-letter state code filter}
\item{max_results}{Max results (default 1000)}
}
\value{tibble: facility_id, facility_name, address, city, state,}
\description{Hospital general information}
