% Generated by build script
\name{cms_nursing_homes}
\alias{cms_nursing_homes}
\title{Nursing home (skilled nursing facility) information}
\usage{
cms_nursing_homes()
}
\arguments{
\item{state}{2-letter state code}
\item{max_results}{Max results (default 1000)}
}
\value{tibble with facility info, ratings, staffing data}
\description{Nursing home (skilled nursing facility) information}
