% Generated by build script
\name{cms_search}
\alias{cms_search}
\title{Search CMS datasets by keyword}
\usage{
cms_search()
}
\arguments{
\item{query}{Search term (case-insensitive, matches title and description)}
}
\value{tibble: identifier, title, description, modified}
\description{Search CMS datasets by keyword}
